+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @voltagent/cli
 
+## 0.1.3
+
+### Patch Changes
+
+- [#33](https://github.com/VoltAgent/voltagent/pull/33) [`3ef2eaa`](https://github.com/VoltAgent/voltagent/commit/3ef2eaa9661e8ecfebf17af56b09af41285d0ca9) Thanks [@kwaa](https://github.com/kwaa)! - Update package.json files:
+
+  - Remove `src` directory from the `files` array.
+  - Add explicit `exports` field for better module resolution.
+
 ## 0.1.1
 
 - 🚀 **Introducing VoltAgent: TypeScript AI Agent Framework!**